Create Pokemon SQL database
This SQL table compiles a list of Pokemon and their stats. The queries will show some examples of fast Pokemon and Pokemon with high attack stats to help build a team. The last query will show the different type combonations on Pokemon to better help prepare for dealing with type matchups.